+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Breaking changes:
   - Removed `Skin::has_cursor_middle` and `Render::has_cursor_middle` ([#2])
+  - Added the fields `RenderOptions::show_strain_graph`, `RenderOptions::show_slider_breaks`, and `RenderOptions::ignore_fail`
 
 - Additions:
   - Implement `Default` for `OrdrClient`